Clean up to continue the tutorial
Before we proceed with the tutorial and implement TicketMonster, we need to
clean up some of the archetype-generated code. The Member management code,
while useful for illustrating the general setup of a Java EE 6 web
application, will not be part of TicketMonster, so we can safely remove
some packages, classes, and resources.